LP 1486254: Convert nested YAML maps
Moved the conform() function from cmd/juju/action into the
common directory for commands so that the system create-env
command can also use it to convert nested maps into the proper
map[string]interface{} format.
