+++ cmd/juju/action/do.go
@@ -15,6 +15,7 @@ import (
 	"launchpad.net/gnuflag"
 
 	"github.com/juju/juju/apiserver/params"
+	"github.com/juju/juju/cmd/juju/common"
 )
 
 var keyRule = regexp.MustCompile("^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$")
@@ -181,7 +182,7 @@ func (c *DoCommand) Run(ctx *cmd.Context) error {
 			return err
 		}
 
-		conformantParams, err := conform(actionParams)
+		conformantParams, err := common.ConformYAML(actionParams)
 		if err != nil {
 			return err
 		}
@@ -211,7 +212,7 @@ func (c *DoCommand) Run(ctx *cmd.Context) error {
 		addValueToMap(keys, cleansedValue, actionParams)
 	}
 
-	conformantParams, err := conform(actionParams)
+	conformantParams, err := common.ConformYAML(actionParams)
 	if err != nil {
 		return err
 	}
